Move tests from tests root namespace to Functional namespace
