+++ tests/Functional/PluginCreator/OutputDirectoryProvider/CurrentOutputDirectoryProviderTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ShopwareCli\tests\PluginCreator\OutputDirectoryProvider;
+namespace ShopwareCli\Tests\Functional\PluginCreator\OutputDirectoryProvider;
 
 use Shopware\PluginCreator\Services\WorkingDirectoryProvider\CurrentOutputDirectoryProvider;
