+++ tests/Functional/PluginCreator/OutputDirectoryProvider/RootDetector/ShopwareRootDetectorTest.php
